Fix bug in validate_num_args
UPDATE: realized I don't need to validate counts for functions that take 0+
args. It's still more semantic the way it is now.

Was not able to allow 0+ arguments previously, it required max to be less than
min to have an unlimited max arg count, which doesn't work when min is 0.
Discovered the usize::MAX constant which is a better solution.

Somehow this brought up a typo in eval -> SpecialForm::Eval handling which only
worked with the incorrect version of validate_num_args somehow.